+++ januscloud/sentinel/main.py
@@ -57,7 +57,8 @@ def do_main(config):
             pingpong_interval=config['janus']['pingpong_interval'],
             statistic_interval=config['janus']['statistic_interval'],
             request_timeout=config['janus']['request_timeout'],
-            hwm_threshold=config['janus']['hwm_threshold']
+            hwm_threshold=config['janus']['hwm_threshold'],
+            admin_secret=config['janus']['admin_secret']
         )
 
         # set up janus_watcher