+++ CHANGES.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
  [v0.2.0]  - unreleased
 ---------------------------------
 
+* Janus-proxy support api_secret authorization
+* Janus-sentinel support admin_secret for sending admin API request
 
 
  [v0.1.0]  - 2020-03-29